apply for slow query list
