+++ ui/lib/apps/Statement/components/StatementsTable.tsx
@@ -1,7 +1,6 @@
 import { usePersistFn } from 'ahooks'
-import React, { useCallback, useEffect, useRef } from 'react'
+import React, { useCallback } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { IDetailsList } from 'office-ui-fabric-react/lib/DetailsList'
 
 import openLink from '@lib/utils/openLink'
 import { CardTable, ICardTableProps } from '@lib/components'
@@ -44,11 +43,6 @@ export default function StatementsTable({ controller, ...restPrpos }: Props) {
 
   const getKey = useCallback((row) => `${row.digest}_${row.schema_name}`, [])
 
-  const tableRef = useRef<IDetailsList>(null)
-  useEffect(() => {
-    tableRef.current?.scrollToIndex(getClickedItemIndex())
-  }, [getClickedItemIndex])
-
   return (
     <CardTable
       {...restPrpos}
@@ -62,7 +56,6 @@ export default function StatementsTable({ controller, ...restPrpos }: Props) {
       visibleColumnKeys={visibleColumnKeys}
       onRowClicked={handleRowClick}
       getKey={getKey}
-      componentRef={tableRef}
       clickedRowIndex={getClickedItemIndex()}
     />
   )
